// packages/data/src/generated/NewTodo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TodoPriority } from "./TodoPriority";

export type NewTodo = { id: string, projectId: string, title: string, description: string | null, priority: TodoPriority, plannedDate: string | null, pinned: boolean, rank: number, };